% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageScatterPlot.R
\name{ImageScatterPlot}
\alias{ImageScatterPlot}
\title{Scatter plot with images at the specified points}
\usage{
ImageScatterPlot(latent, observed, observed.dim, bins = 256,
  col.palette = viridis::viridis, image.density = 10, num.attempts = 100,
  interpolate = F, ...)
}
\arguments{
\item{latent}{An n-by-2 matrix specifying where points in observed should be
plotted.}

\item{observed}{An n-by-q matrix of image data to be plotted at the
corresponding points in \code{latent}.}

\item{observed.dim}{A numeric vector of length 2 giving the image dimensions
for each row in \code{observed}. \code{prod(observed.dim) ==
ncol(observed)} must be true.}

\item{bins}{The number of colors to generate from col.palette.}

\item{col.palette}{A color palette function, such as rainbow or heat.colors, or a vector of colors of length \code{bins}.
Default is \link[viridis]{viridis}.}

\item{image.density}{How dense should the plotted images be. Higher
\code{image.density} means smaller images.}

\item{num.attempts}{The number of non-overlapping plots to generate. The one
which plots the most points will be chosen for the output plot.}

\item{...}{Any additional parameters to be passed to \link{plot}.}
}
\value{
A logical vector indicating which rows have been plotted (returned
  invisibly).
}
\description{
Create a scatter plot of image data. Images are plotted in random order, and
are only plotted if they do not overlap with an already-plotted image. See
the examples section for an example using MNIST data.
}
\examples{
#Plot PCA of the MNIST data
mnist <- ImageScatterPlot::mnist
Z <- prcomp(mnist)$x[,1:2]
ImageScatterPlot(Z,
                 mnist,
                 rep(28,2),
                 col.palette = grey.colors(256, start=1, end=0),
                 num.attempts = 10,
                 image.density=15,
                 interpolate=T,
                 main="First two principal components of MNIST handwritten digits")
}
